% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_MR_analysis.R
\name{organize_MR_analysis}
\alias{organize_MR_analysis}
\title{Title}
\usage{
organize_MR_analysis(
  create = c("AUTO", "MANUAL", "Full", "BACTERIAL_RESP", "SDA")
)
}
\arguments{
\item{create}{folders to create to organise local export datafiles}
}
\value{
None, saved folder
}
\description{
Title
}
\examples{

\dontrun{
organize_MR_analysis(create="none")
}
}
