% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRrepeat.R
\name{RMRrepeat}
\alias{RMRrepeat}
\title{Title}
\usage{
RMRrepeat(
  data,
  cycle_start,
  cycle_end,
  chop_start,
  chop_end,
  N_Ch = 4,
  inventory_data = NA,
  flush_plot = FALSE,
  path = ".",
  date_format = c("m/d/y", "d/m/y", "y-m-d"),
  plot_temp = FALSE
)
}
\arguments{
\item{data}{The name of the raw input file .csv, character.}

\item{cycle_start}{The stat time (min) of auto cycle. This indicates the start of the measurement (relative to min 0).}

\item{cycle_end}{The time (min) of auto cycle; end of the measurement (relative to min 0)}

\item{chop_start}{The time (min) to be chopped off at the start of each measurement cycle}

\item{chop_end}{The time (min) to be chopped off at the end of each measurement cycle}

\item{N_Ch}{The number of channels of the oxygen meter. It must be either 4 or 8. 8-Channel Firesting has 4 temperature and 4 oxygen sensors}

\item{inventory_data}{The inventory file (.csv file; must be in the same working directory as the indicated data file)}

\item{flush_plot}{Plotting of additional figures where full cycle (measure:flush) is presented.}

\item{path}{Local directory to save output files. Two options, either all files will be saved in the current working directory, or the other option requires the folders created by organize_MR_analysis}

\item{date_format}{The date format used in the original FireSting files. Must specify one of the following: c("m/d/y","d/m/y","y-m-d")}

\item{plot_temp}{Logical argument. Indicates whether or not temperature trends for each cycle will be plotted and saved    TRUE}
}
\value{
The output from \code{\link{print}}
}
\description{
Title
}
