% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_SDA.R
\name{SDA}
\alias{SDA}
\title{Title}
\usage{
SDA(
  AnimalID,
  BW.animal,
  resp.V,
  r2_threshold_smr,
  sda_threshold_level,
  scaling_exponent_mmr = 1,
  scaling_exponent_smr = 1,
  date_format = c("\%m/\%d/\%Y \%H:\%M:\%S", "GMT"),
  data.SDA = NULL,
  analyzed_MR = NULL,
  SMR_calc = TRUE,
  SMR_vals = c(NA, NA, NA, NA),
  drop_ch = NULL,
  N_Ch = 4,
  end_SDA_Ch = NA,
  MLND = TRUE,
  background_prior = NULL,
  background_post = NULL,
  background_slope = NULL,
  background.V = NULL,
  match_background_Ch = FALSE,
  background_linear_gr = FALSE,
  background_exp_gr = FALSE,
  local_path = FALSE,
  handling_delay = 0,
  begin_smr_hr_zero = FALSE
)
}
\arguments{
\item{begin_smr_hr_zero}{}
}
\value{
The output from \code{\link{print}}
}
\description{
Title
}
