% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMR_SMR_AS_EPOC.R
\name{MMR_SMR_AS_EPOC}
\alias{MMR_SMR_AS_EPOC}
\title{estimates indivudal specific metabolic rates and recovery}
\usage{
MMR_SMR_AS_EPOC(
  data.MMR = NULL,
  data.SMR = NULL,
  AnimalID,
  BW.animal,
  resp.V,
  r2_threshold_smr,
  r2_threshold_mmr,
  min_length_mmr,
  scaling_exponent_mmr = 1,
  scaling_exponent_smr = 1,
  common_mass = 1,
  mo2_val_for_calc = "mo2_1kg",
  plot_smr_quantile = 15,
  date_format = c("\%Y-\%m-\%d \%H:\%M:\%S", "GMT"),
  N_Ch = 4,
  drop_ch = NULL,
  MLND = TRUE,
  verbose.MLND = FALSE,
  calc_EPOC = TRUE,
  epoc_threshold = 1,
  recovMMR_threshold = 0.5,
  end_EPOC_Ch = c(NA, NA, NA, NA),
  background_prior = NULL,
  background_post = NULL,
  background_slope = NULL,
  background_slope_mmr = NULL,
  background.V = NULL,
  background_gr = NULL,
  match_background_Ch = FALSE,
  mmr_background = "trial_mean",
  local_path = TRUE
)
}
\arguments{
\item{data.MMR}{The name of the MMR data file (“…analyzed.csv”; a character string); an output file from the MMR function.}

\item{data.SMR}{The name of the SMR data file (“…analyzed.csv”; a character string); an output file from the SMR function.}

\item{AnimalID}{Indicates individual ID; must be a vector of 4 characters. When missing, enter "NA"}

\item{BW.animal}{Indicates individual mass; must be a vector of 4 characters. When missing, enter "0"}

\item{resp.V}{Indicates the volume (L) of respirometery chambers; must be a vector of 4 numbers (e.g., c(1, 1, 1, 1), for four 1-L respirometers)}

\item{r2_threshold_smr}{R2 threshold for SMR, measurements below the threshold are excluded}

\item{r2_threshold_mmr}{R2 threshold for MMR, measurements below the threshold are excluded}

\item{min_length_mmr}{The duration of MMR steepest slope measurement; 180, 120, 90, 60 seconds (s)}

\item{scaling_exponent_mmr}{Body mass scaling exponent to correct MMR values for body size. MR=aBM^b (MR = metabolic rate, BW = body mass, a = scaling coefficient [the intercept], and b = scaling exponent [the power term])}

\item{scaling_exponent_smr}{Body mass scaling exponent to correct SMR values for body size. MR=aBM^b (MR = metabolic rate, BW = body mass, a = scaling coefficient [the intercept], and b = scaling exponent [the power term])}

\item{common_mass}{Metabolic performances are often calculated per unit mass. Use this argument to define what the standardized mass should be. (default is MO2 mgO2kg-1^ min-1^, a common mass of 1 kg). Units = kg}

\item{mo2_val_for_calc}{Units of metabolic rates represented on export figures. mo2_common_mass_kg (standardized MO2 to 1 kg fish using scaling exponents provided; mg O2 min-1 kg-1), mo2_individual_kg (MO2 of the whole individual, mgO2 min-1)}

\item{plot_smr_quantile}{Indicating what percentile lower MO2 values to plot; must be either 10, 15, or 20}

\item{date_format}{The date format used in the original data files. Argument is passed to strptime. default is c("\%m/\%d/\%Y \%H:\%M:\%S", "GMT").}

\item{N_Ch}{The number of channels of the oxygen meter. It must be either 4 or 8 (8 channem meter has 4 oxygen probes and 4 temperature probes)}

\item{drop_ch}{Indicates which channel is dropped or entirely excluded from the analysis. Must be a numerical vector, e.g., c(1,3)}

\item{MLND}{Logical argument. If TRUE, SMR is estimated also using Mean Lowest Normal Distribution analysis. More details in Chabot et al 2016}

\item{verbose.MLND}{From MLND: A logical controlling if a text progress bar from MLND is displayed during the fitting procedure. (see 'verbose' in mclust package functions).}

\item{calc_EPOC}{Logical. If both SMR and MMR files are provided, indicate whether or not to evaluate recovery (i.e. EPOC, hourcly recovery, etc.)}

\item{epoc_threshold}{Indicates the threshold relevant to an individual’s SMR to calculate the end time of recovery (the time at which metabolic rate has returned to epoc_threshold). The default is the SMR level (1; At 100 percent SMR). To use 120 percent SMR as a recovery threshold, enter epoc_threshold = 1.2.}

\item{recovMMR_threshold}{Short-term recovery threshold relevant to individual’s MMR. Indicates the (percent) level of MMR to calculate recovery indices. The default indicates 50 percent MMR, e.g., the time it takes to recover to 50 percent MMR (default). Enter 0.3 for 30 percent MMR, 0.8 for 80 percent MMR)}

\item{end_EPOC_Ch}{Manually assigned end time (min) of full recovery (EPOC). The time can be assigned for each channel independently; use a numerical vector of 4 variables, one for each channel (e.g., c(120, 120, NA, 180), for 2 h, 2h, SMR level, and 3 h EPOC end times, respectively)}

\item{background_prior}{The name of the analyzed background “…analyzed.csv” file, an output file from the SMR function (respiration in an empty respirometer measured before the respirometry trial).}

\item{background_post}{The same as background_prior, only post respirometry trial}

\item{background_slope}{Manually assigned background slope used to correct metabolic rate in all individuals. Provide numeric value in mgO2 L-1 h-1}

\item{background_slope_mmr}{Manually assigned background slope used to correct MMR in all individuals. Provide numeric value in mgO2 L-1 h-1}

\item{background.V}{Manually assigned respirometer volumes (L). A vector with 4 numeric variables, one for each channel.}

\item{background_gr}{Specify whether to assume that bacterial growth (thus respiration rates) changed linearly or in exponentially across the duration of the respirometry trial. Must specify either "linear" or "exp": metabolic rate values across the given trial are corrected using the estimated background values from the indicated growth curve. Both background_prior and background_post must be provided to enable this.}

\item{match_background_Ch}{Logical. If TRUE, the background respiration is estimated and applied channel-specific. The background_prior and background_post are used to estimate background respiration specific to each channel (or individual), which then is used to correct each individual’s MO2 independently. By default, the mean background respiration rate from all channels is calculated and applied to correct all individual’s MO2}

\item{mmr_background}{Specifies what background value should be used to correct MMR value. Options: i) "back_prior" takes background respiration rate value estimated from the background_prior file only, ii)"trial_mean" takes background respiration rate value that is applied to the entire trial (MMR and SMR, indistinguishable), iii) "back_gr", and iv) a defined respiration rate value mgO2 L-1 h-1.}

\item{local_path}{Logical. If TRUE (default) all returned files will be saved in the local working directory.}
}
\value{
The output from \code{\link{print}}
}
\description{
Estimates aerobic metabolic performances in animals: MMR, SMR, EPOC (recovery), and Absolute Aerobic Scope. Relies on outputs from SMR and MMR functions
}
