% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_raw_data_read.R
\name{sample_raw_data_read}
\alias{sample_raw_data_read}
\title{Read text file and then start the data analysis sequence}
\usage{
sample_raw_data_read(path)
}
\arguments{
\item{path}{}
}
\value{
a \code{tibble}
}
\description{
Read text file and then start the data analysis sequence
}
\examples{
txt = system.file("extdata", "raw_txt_data_MMR.txt", package = "respMR")
sample_raw_data_read(txt)
}
