% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR.R
\name{SMR}
\alias{SMR}
\title{Title}
\usage{
SMR(
  data,
  cycle_start,
  cycle_end,
  chop_start,
  chop_end,
  N_Ch = 4,
  date_format = c("\%m/\%d/\%Y \%H:\%M:\%S", "GMT"),
  inventory_data = NULL,
  flush_plot = NULL,
  local_path = TRUE,
  plot_temp = FALSE,
  background_data = FALSE,
  sda_data = FALSE
)
}
\arguments{
\item{data}{The name of the raw input file .csv, character.}

\item{cycle_start}{The stat time (min) of auto cycle. This indicates the start of the measurement (relative to min 0).}

\item{cycle_end}{The time (min) of auto cycle; end of the measurement (relative to min 0)}

\item{chop_start}{The time (min) to be chopped off at the start of each measurement cycle}

\item{chop_end}{The time (min) to be chopped off at the end of each measurement cycle}

\item{N_Ch}{The number of channels of the oxygen meter. It must be either 4 or 8. 8-Channel Firesting has 4 temperature and 4 oxygen sensors}

\item{date_format}{The date format used in the original FireSting files. Must specify one of the following: c("m/d/y","d/m/y","y-m-d")}

\item{inventory_data}{The inventory file (.csv file), default = NULL}

\item{flush_plot}{Plotting of additional figures where full cycle (measure:flush) is presented.}

\item{local_path}{Logical. If TRUE (default) all returned files will be saved in the local working directory.}

\item{plot_temp}{Logical argument. Indicates whether or not temperature trends for each cycle will be plotted and saved    TRUE}
}
\value{
The output from \code{\link{print}}
}
\description{
Title
}
