% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFileConvert.R
\name{textFileConvert}
\alias{textFileConvert}
\title{Title}
\usage{
textFileConvert(
  txt_file,
  nrowSkip,
  N_Ch = 4,
  local_path = TRUE,
  exclude_first_measurement_s = 0
)
}
\arguments{
\item{txt_file}{The name of the original “.txt” file from FireSting}

\item{nrowSkip}{The number of rows to skip; these rows in raw txt file often contain calibration information, the IDs of the probes and other user defined settings}

\item{N_Ch}{The number of FireSting channels. Options include 2, 4, 8. If a 2-channel FireSting was used, this argument could be ignored, or enter 4}

\item{local_path}{Logical. If TRUE (default) all returned files will be saved in the local working directory.}

\item{exclude_first_measurement_s}{The number measurement point to be excluded from the beginning of the file (in addition to the nrowSkip argument)}
}
\value{
The output from \code{\link{print}}
}
\description{
Title
}
